+++ src/js/constants/MONTHS.js
@@ -1,0 +1,1 @@
+export const MONTHS = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
