+++ src/js/events/inputBar/index.js
@@ -1,0 +1,7 @@
+import { $, $All } from "../../utils/dom.js";
+import { priceInput } from "./priceInput.js";
+
+
+export const inputBarEventHandler = () => {
+    $('#price-input').addEventListener("keyup",priceInput)
+};
