+++ src/js/utils/dom.js
@@ -1,1 +1,4 @@
-export const $ = (selector) => document.querySelector(selector);
+const $ = (selector) => document.querySelector(selector);
+const $All = (selector) => document.querySelectorAll(selector);
+
+export { $, $All };
